requirements.txt
gradio>=4.44.0
deep-translator>=1.11.4
openai-whisper>=20231117
gTTS>=2.5.1
pypdf>=4.3.0
docx2txt>=0.8
